+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Upcoming
 
 - Add `includedHosts`, `includedURLs`, `excludedHosts`, and `excludedURLs` to `NetworkLogger/Configuration`. By default, they support simple wildcards, e.g. `*.example.com`, but you can also enable full regex using another new configuration options: `isRegexEnabled`.
-- Add `excludedHeaders` to `NetworkLogger/Configuration` for redacting sentitive information from logged HTTP headers 
+- Add `excludedHeaders`, `excludedDataFields` to `NetworkLogger/Configuration` for redacting sentitive information from logged HTTP headers 
 - Add a new convenience initializer to `NetworkLogger` with `configure` trailing closure
 - Make `LoggerStore.Event` frozen
 - Add cookies section and details screens to the request summary - [#124](https://github.com/kean/Pulse/pull/124)